Delete the Author site in "Specifies the page to be used for memory release.".
Evidently, the tab doesn't seem release the memory.
for example,
"in the extension" -> using the memory of purged tab is 22M(20M = Extension, 2M = purged tab).
"Author size" -> the tab doesn't seem release the memory.
"Assigment file" -> Only the first time, purging tab. But this problem is cause the history of the tab.