Changed: the behavior of the toggle temporary exclusion list.
I have implemented the function to add the hostname
to the temporary exclusion list.
Also I have implemented a method to remove it.

When the popup menu is displaying,
If the user is pressing the Ctrl key,
add the url to the temporary exclusion list,
or display the item that delete.
When the Ctrl key isn't pressing, display the item for hostname to add.

The behavior of when adds to the list.
Add a hostname, then delete a hostname = delete it,
Add a hostname, then delete a url      = doesn't delete it.
Add a URL,      then delete a hostname = delete it.
Add a URL,      then delete a url      = delete it.